% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{es}
\alias{es}
\title{Calculate the 'enrichment score' for gene-set amongst ranking/correlations}
\usage{
es(N, S, p = 1, r = N:1/N)
}
\arguments{
\item{N}{Integer value}

\item{S}{Ranks of gene set}

\item{p}{Weighting of ranking/correlations}

\item{r}{Rank/correlation scores}
}
\value{
Numeric value - enrichment score
}
\description{
Calculate the 'enrichment score' for gene-set amongst ranking/correlations
}

