% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interface.R
\name{get_ontological_gene_sets}
\alias{get_ontological_gene_sets}
\title{Create list of gene sets defined by ontological annotation}
\usage{
get_ontological_gene_sets(ontology, gene, term, min_genes = 1,
  max_genes = 500)
}
\arguments{
\item{ontology}{\code{ontology_index} object.}

\item{gene}{Character vector of genes.}

\item{term}{Character vector of term IDs annotated to corresponding genes.}

\item{min_genes}{Minimum number of genes in gene sets.}

\item{max_genes}{Maximum number of genes in gene sets.}
}
\value{
List of character vectors of term IDs.
}
\description{
Create list of gene sets defined by ontological annotation
}
