% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interface.R
\docType{data}
\name{GO_gene_sets}
\alias{GO_gene_sets}
\title{GO term gene sets}
\format{List of character vectors of genes per GO term, and named by term ID.}
\description{
List of gene sets annotated by each GO term
}
\details{
Based on gene-GO term annotations downloaded from \code{geneontology.org}. Only contains gene sets for terms with up to 500 genes.
}
