\name{gsEasy-package}
\alias{gsEasy-package}
\alias{gsEasy}
\docType{package}
\title{
Gene Set Enrichment Analysis in R
}
\description{
R-interface to C++ implementation of the rank/score based GSEA test described by Subramanian et al 2005.
}
\details{
\tabular{ll}{
Package: \tab gsEasy\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2016-01-11\cr
License: \tab GPL >= 2\cr
}
%~~ An overview of how to use the package, including the most important ~~
%~~ functions ~~
}
\author{
Daniel Greene <dg333@cam.ac.uk>

Maintainer: Daniel Greene <dg333@cam.ac.uk>
}
\references{
Subramanian, A, Tamayo, P, Mootha, VK, Mukherjee, S, Ebert, BL, Gillette, MA, Paulovich, A, Pomeroy, SL, Golub, TR, Lander, ES, Mesirov, JP (2005). Gene set enrichment analysis: a knowledge-based approach for interpreting genome-wide expression profiles. Proc. Natl. Acad. Sci. U.S.A., 102, 43:15545-50, doi: 10.1073/pnas.0506580102.

Ashburner et al. Gene ontology: tool for the unification of biology (2000) Nat Genet 25(1):25-9
}

